:construction: still figuring out best way to handle gfs
how to handle gfs with hidden_external_contribs that will end up being not hidden? maybe a way to do this is to tell the user "run the tool twice"? i don't like that. Maybe get all the descendants of a study and then work through them to see if they should or should not be hidden in the end....